Adapted bug fix for flake8

+++ scrapers/mexico.py
@@ -95,12 +95,8 @@ class Mexico:
 
     def __del__(self):
         self.driver.quit()
-
-        ## TODO: possible errors in this line
-        try:
+        if os.path.isdir(self.downloads_dir):
             shutil.rmtree(self.downloads_dir)
-        except:
-            pass
 
     def __manual_click(self, element):
         WebDriverWait(self.driver, 10).until(
